feat(edge): roll out static OG manifest and restore example previews
