+++ scripts/edge-validation-utils.mjs
@@ -23,25 +23,59 @@ export const parseEdgeFunctionEntries = (toml) => {
 export const findCatchAllEdgeEntries = (entries) =>
   entries.filter((entry) => entry.path.trim() === "/*");
 
-const SITE_OG_META_ALLOWED_PATHS = new Set([
+const SITE_OG_META_LOCALES = ["es", "de", "fr", "pt", "ru", "it", "pl", "ko"];
+
+const SITE_OG_META_BASE_ALLOWED_PATHS = [
+  "/",
+  "/features",
+  "/inspirations",
+  "/inspirations/*",
+  "/updates",
   "/blog",
   "/blog/*",
-  "/es/blog",
-  "/es/blog/*",
-  "/de/blog",
-  "/de/blog/*",
-  "/fr/blog",
-  "/fr/blog/*",
-  "/pt/blog",
-  "/pt/blog/*",
-  "/ru/blog",
-  "/ru/blog/*",
-  "/it/blog",
-  "/it/blog/*",
-  "/pl/blog",
-  "/pl/blog/*",
-  "/ko/blog",
-  "/ko/blog/*",
+  "/pricing",
+  "/faq",
+  "/share-unavailable",
+  "/login",
+  "/contact",
+  "/imprint",
+  "/privacy",
+  "/terms",
+  "/cookies",
+  "/create-trip",
+  "/example/*",
+];
+
+const SITE_OG_META_LOCALIZED_ALLOWED_PATHS = [
+  "/",
+  "/features",
+  "/inspirations",
+  "/inspirations/*",
+  "/updates",
+  "/blog",
+  "/blog/*",
+  "/pricing",
+  "/faq",
+  "/share-unavailable",
+  "/login",
+  "/contact",
+  "/imprint",
+  "/privacy",
+  "/terms",
+  "/cookies",
+  "/create-trip",
+];
+
+const localizePath = (path, locale) => {
+  if (path === "/") return `/${locale}`;
+  return `/${locale}${path}`;
+};
+
+const SITE_OG_META_ALLOWED_PATHS = new Set([
+  ...SITE_OG_META_BASE_ALLOWED_PATHS,
+  ...SITE_OG_META_LOCALES.flatMap((locale) =>
+    SITE_OG_META_LOCALIZED_ALLOWED_PATHS.map((path) => localizePath(path, locale))
+  ),
 ]);
 
 export const findSiteOgMetaScopeViolations = (entries) =>
@@ -50,5 +84,5 @@ export const findSiteOgMetaScopeViolations = (entries) =>
     .filter((entry) => !SITE_OG_META_ALLOWED_PATHS.has(entry.path.trim()))
     .map((entry) => ({
       ...entry,
-      reason: `site-og-meta must only be mapped to blog routes. Found disallowed path "${entry.path}".`,
+      reason: `site-og-meta must only be mapped to approved static/example route allowlists. Found disallowed path "${entry.path}".`,
     }));
